backend/requirements.txt
fastapi
uvicorn
transformers
torch --index-url https://download.pytorch.org/whl/cpu
google-cloud-texttospeech
google-genai
huggingface_hub
Pillow